Feat: Rename directory
