+++ src/me/day03/assignment01_review/Practice04.java
@@ -1,0 +1,14 @@
+package me.day03.assignment01_review;
+
+public class Practice04 {
+    public static void main(String[] args) {
+
+        User user = new User();
+        user.setOriginalMoney(Utils.input("얼마 가지고 있니? "));
+
+        Kiosk kiosk = Kiosk.getInstance();
+        kiosk.order(user);
+
+    }
+
+}
