+++ src/me/day03/assignment01_review/Utils.java
@@ -1,0 +1,13 @@
+package me.day03.assignment01_review;
+
+import java.util.Scanner;
+
+public class Utils {
+    private static Scanner scanner = new Scanner(System.in);
+
+    public static int input(String message) {
+        System.out.print(message);
+        return scanner.nextInt();
+    }
+
+}
