+++ src/me/day05/assignment02_review/practice03/Main.java
@@ -1,0 +1,8 @@
+package me.day05.assignment02_review.practice03;
+
+public class Main {
+    public static void main(String[] args) {
+        Clazz cls = new Clazz();
+        System.out.println(cls);
+    }
+}
